Fixed setGenerate in regtest mode
When calling setGenerate while running bitcoind in regtest mode,
setGenerate is supposed to parse a [HexString] rather than Nil in some cases.
In regtest mode the setgenerate API call doesn't take a number of processors,
but rather a number of hashes to generate, and than proceeds to return these hashes.